Image Subtitles
By adding few lines of code we can accomplish to have titles + subtitles
texts for your images. The modified files create the "subtitle text"
field and the "subtitle position" field in the SiteOrigin Image widget
backend.

I tested for previous version localy and now I've made the changes in
the updated plugin (Version 1.5.11) and working fine.

This is a minor change that could be easily integrated for the new
version.