Prevent pagination inline-block element spaces
Resolves #1005.

Pagination spacing relies on the extra space created between inline-block elements. If a user enables HTML optimization in a plugin like AO those spaces will be removed when the HTML is minimized and the new lines are removed.

It isn’t necessary to reset the font size for the list items but I did it anyway.

According to this page https://stackoverflow.com/questions/5078239/how-do-i-remove-the-space-between-inline-block-elements the method I’ve used is compatible as follows:

This works in recent versions of all modern browsers. It works in IE8. It does not work in Safari 5, but it does work in Safari 6. Safari 5 is nearly a dead browser (0.33%, August 2015).

Another option would be to format our HTML so that each list item for the pagination is on the same line.