Post Carousel: Side Step Slick Navigation
A custom navigation is used due to a Slick limitation that prevents the slide from stopping the slide from changing and wanting to remain consistent with the previous carousel.

https://github.com/kenwheeler/slick/pull/2104

I didn't previously think this was necessary as stopImmediatePropagation() was working. In hindsight, that's obviously unreliablein this instance as it relies on this code firing before Slicks which isn't always the case.