Carousel: Remove Generic CSS
Merged existing generic CSS with post carousel and (in a different PR) the Anything Carousel. This file is useful for keeping things DRY, but there's just not as many overlaps as I had thought there was going to be.

I also prefixed the post carousel CSS to prevent potential conflicts with other SO Carousel widgets.