Blog Widget: Standard paginate_links changes
If ajax is enabled, the current must be set to an invalid number to prevent the correct current from being set a span.

`show_all` must also be enabled to ensure all items are output.
`prev_next` can't be forced, so it has to be disabled when Ajax is enabled.