Carousel: Container Navigation
This commit adds an optional parameter to the carousel that allows for the carousel dots to be added to the `.sow-carousel-nav` class.

This change allows for the carousel dots and arrow navigation to appear on the same line.

The LESS change allows for the navigation arrows to work with this structure change.